Renamed folders and file directories, updated comments from YAML
- Renamed various folders and file directories for better organization
- Updated comments in YAML files to reflect the new structure
