+++ .github/workflows/ci.yml
@@ -60,51 +60,6 @@ jobs:
           SONAR_HOST_URL: ${{ secrets.SONAR_HOST_URL }}
           DOTNET_ROOT: /home/runner/.dotnet
 
-#  update_readme:
-#    runs-on: ubuntu-latest
-#    needs: sonar_scan
-#
-#    steps:
-#      - name: Checkout code
-#        uses: actions/checkout@v3
-#
-#      - name: Fetch SonarQube metrics
-#        id: fetch_sonar_metrics
-#        run: |
-#          curl -u ${{ secrets.SONAR_TOKEN }}: \
-#            "${{ secrets.SONAR_HOST_URL }}/api/measures/component?component=${{ secrets.SONAR_PROJECT_KEY }}&metricKeys=coverage,bugs,vulnerabilities,code_smells" \
-#            -o sonar_metrics.json
-#
-#      - name: Update README.md
-#        run: |
-#          coverage=$(jq -r '.component.measures[] | select(.metric == "coverage") | .value' sonar_metrics.json)
-#          bugs=$(jq -r '.component.measures[] | select(.metric == "bugs") | .value' sonar_metrics.json)
-#          vulnerabilities=$(jq -r '.component.measures[] | select(.metric == "vulnerabilities") | .value' sonar_metrics.json)
-#          code_smells=$(jq -r '.component.measures[] | select(.metric == "code_smells") | .value' sonar_metrics.json)
-
-#          sed -i "s/<!-- sonar_coverage -->.*/<!-- sonar_coverage -->Coverage: $coverage%/" README.md
-#          sed -i "s/<!-- sonar_bugs -->.*/<!-- sonar_bugs -->Bugs: $bugs/" README.md
-#          sed -i "s/<!-- sonar_vulnerabilities -->.*/<!-- sonar_vulnerabilities -->Vulnerabilities: $vulnerabilities/" README.md
-#          sed -i "s/<!-- sonar_code_smells -->.*/<!-- sonar_code_smells -->Code Smells: $code_smells/" README.md
-
-#      - name: Check for changes
-#        run: |
-#          if git diff --quiet; then
-#            echo "No changes to commit"
-#            exit 0
-#          fi
-#          
-#      - name: Commit and push changes
-#        env:
-#          GH_PAT: ${{ secrets.GH_PAT }}
-#        run: |
-#          git config --global user.name 'github-actions[bot]'
-#          git config --global user.email 'github-actions[bot]@users.noreply.github.com'
-#          git add README.md
-#          git add sonar_metrics.json
-#          git commit -m "Update SonarQube metrics in README.md and sonar_metrics.json"
-#          git push https://x-access-token:${GH_PAT}@github.com/carlosbrunetti/Respify.git HEAD:${{ github.ref }}
-#  
   publish_nuget:
     runs-on: ubuntu-latest
     needs: sonar_scan
