+++ Respify.Tests/unit/classes/NonPaginatedResponseTests.cs
@@ -1,24 +1,20 @@
 using Respify.Tests.helpers.@class;
 using Respify.Tests.helpers.generator;
 
-namespace Respify.Tests.unity.classes
+namespace Respify.Tests.unit.classes
 {
-    public class PaginatedResponseTests
+    public class NonPaginatedResponseTests
     {
         [Fact]
         public void GetProperties_ShouldReturnCorrectValues()
         {
             // Arrange
             var cars = new CarGenerator().GenerateCarList(10);
-            var response = new PaginatedResponse<List<Car>>(cars, cars.Count(), 1, 10, "year", "asc");
-
+            var response = new NonPaginatedResponse<List<Car>>(cars, cars.Count);
+                
             // Act & Assert
             Assert.Equal(cars, response.Items);
             Assert.Equal(10, response.Count);
-            Assert.Equal(1, response.PageNumber);
-            Assert.Equal(10, response.PageSize);
-            Assert.Equal("year", response.OrderBy);
-            Assert.Equal("asc", response.SortBy);
         }
     }
 }
