+++ Respify/interfaces/IPaginatedResponse.cs
